added users and stats, closeAll

+++ Auth_web/api/admin_utils/users/delete.php
@@ -2,8 +2,12 @@
 /** @noinspection PhpUndefinedVariableInspection */
 include_once $_SERVER['DOCUMENT_ROOT'] . '/api/config/CodesAndMessages.php';
 
-if($database->delete_user($user_data['user_id'])){
-    $message = 'Успешно!';
+if(isset($user_data['user_id'])) {
+    if ($database->delete_user($user_data['user_id'])) {
+        $message = 'Успешно!';
+    } else {
+        $error_code = CodesAndMessages::DB_ERROR;
+    }
 } else {
-    $error_code = CodesAndMessages::DB_ERROR;
+    $error_code = CodesAndMessages::WRONG_REQUEST_PARAMS;
 }
